move test into general workflow test
